\name{alpha_rand}
\alias{alpha_rand}
\title{Generating alpha for GMM distribution}

\description{
Function generating alpha values of mixture distribution.
}

\usage{
alpha_rand(
      n,
      m
)
}

\arguments{
\item{n}{Number of alpha sets}

\item{m}{Number of alphas}
}


\examples{
alpha_rand(1,4)
}

}
