% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binned}
\alias{binned}
\title{Binned data of mass spectrometry}
\format{
A matrix of X and Y (in histogram)
}
\usage{
data(binned)
}
\description{
This data set is part of mass spectrometry measurements. First column represent X values. Second column represent counts of X.
}
\keyword{datasets}
