% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{coords_to_img}
\alias{coords_to_img}
\title{2D plot support}
\usage{
coords_to_img(X, Y)
}
\arguments{
\item{X}{matrix of 2D GMM data.}

\item{Y}{y coordinates of histogram.}
}
\description{
Supporting function for 2D GMM ploting
}
