% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2D}
\alias{data_2D}
\title{Data of 2D mixture}
\format{
Lists of measurements of 2D data
}
\usage{
data(data_2D)
}
\description{
This data set contain 336 images transformed into data list with X and Y coordinates and counts
}
