% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binned.R
\docType{data}
\name{data_binned}
\alias{data_binned}
\title{Binned data of mass spectrometry}
\format{
A matrix of X and Y
}
\usage{
data(data_binned)
}
\description{
This data set is part of mass spectrometry measurements. First column represent X values. Second column represent counts of X.
}
\keyword{datasets}
