% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{diag_init_2D}
\alias{diag_init_2D}
\title{Diagonal initialization of 2D GMM}
\usage{
diag_init_2D(X, KS)
}
\arguments{
\item{X}{Matrix of 2D GMM data.}

\item{KS}{Number of components.}
}
\value{
Initial values for EM.
}
\description{
Function for generating initial conditions of 2D GMM model by diagonal.
}
\keyword{internal}
