% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dsupport.R
\name{dyn_pr_split_w}
\alias{dyn_pr_split_w}
\title{Dynamic programming split}
\usage{
dyn_pr_split_w(xhist, yhist, K, aux_mx, s_corr)
}
\arguments{
\item{xhist}{x coordinates of histogram.}

\item{yhist}{y coordinates of histogram.}

\item{K}{Number of components.}

\item{aux_mx}{Return of function \code{dyn_pr_split_w_aux}.}
}
\value{
List of parameters.
}
\description{
Supporting function for dynamic programming in 1D decomposition.
}
\keyword{internal}
