% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dplots.R
\name{ellips2D}
\alias{ellips2D}
\title{Support function for plot 2D GMM}
\usage{
ellips2D(center, covariance, cov_type)
}
\arguments{
\item{center}{Means of decomposition}

\item{covariance}{Covariances of each component}

\item{cov_type}{NEED explenation!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!}
}
\description{
Function for managing ellipses in 2D GMM plot
}
