% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Data of 1D mixed-normal distributions}
\format{
A vector containing 1500 observations
}
\source{
Randomly generated data
}
\usage{
data(example)
}
\description{
This data set was randomly drown for 6 components GMM. The parameters of distributions are as follow:\cr
means <- c(-14.56, -14.16, -11.80, -8.77, -2.89, 2.31);\cr
sigma <- c(2.06, 4.49, 4.42, 2.39, 3.92, 1.36);\cr
alpha <- c(0.2012, 0.2898, 0.0334, 0.0092, 0.4278, 0.0384)
}
\keyword{datasets}
