% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2D}
\alias{example2D}
\title{Data of 2D mixed-normal distributions}
\format{
\code{data.frame} of X and Y coordinates and counts
}
\usage{
data(example2D)
}
\description{
This data set contain translated image information into X and Y coordinates and count for each pair X and Y.
}
\keyword{datasets}
