% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2D_2}
\alias{example2D_2}
\title{Data of 2D mixture - 2}
\format{
\code{data.frame} of X and Y coordinates and counts
}
\usage{
data(example2D_2)
}
\description{
This data set contain translated image information into X and Y coordinates and count for each pair X nad Y.
}
\keyword{datasets}
