% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dthresholds.R
\name{find_class_2D}
\alias{find_class_2D}
\title{Class assignment for 2D Gaussian Mixture Model data}
\usage{
find_class_2D(X, gmm)
}
\arguments{
\item{X}{matrix of data to decompose by GMM.}

\item{gmm}{results of \code{\link{gaussian_mixture_2D}} decomposition}
}
\value{
Return a vector of cluster assignment of each point in X matrix.
}
\description{
Function which assign each point of 2D matrix data to a cluster by maximum probability.
}
