% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_merge.R
\name{gmm_merge}
\alias{gmm_merge}
\title{Merging of overlapping components}
\usage{
gmm_merge(GModel, sigmas.dev = 1.5)
}
\arguments{
\item{GModel}{A \code{data.frame} of model component parameters - the rows are, in turn:  mean values (mu), standard deviations (sigma)
and weights (alpha) and each column corresponds to one component. Output of \code{EM_iter}}

\item{sigmas.dev}{Number of sigmas defining the distance to merge the overlapping components of GMM.
By default it is \code{sigma.dev = 2.5}}
}
\value{
Function returns a \code{list} which contains: \describe{
  \item{model}{\code{data.frame} of the GMM decomposition parameters after merging}
  \item{KS}{number of components after merge}
}
}
\description{
Function merges distributions of overlapping components at distances defined in the argument \code{sigmas.dev}.
}
\examples{
GModel<-data.frame(alpha=c(0.45,0.5,0.05),
                  mu=c(-14,-2,5),
                  sigma=c(2,4,1.5))
gmm_merge(GModel, sigmas.dev=5)


}
\seealso{
\code{\link{runGMM}} and \code{\link{EM_iter}}
}
