% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{rand_init_2D}
\alias{rand_init_2D}
\title{Random initialization of 2D GMM.}
\usage{
rand_init_2D(data, KS)
}
\arguments{
\item{KS}{Number of components.}

\item{X}{x coordinates of histogram.}
}
\value{
List of parameters.
}
\description{
Function for generating random initial conditions of 2D GMM model.
}
