% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_rand.R
\name{alpha_rand}
\alias{alpha_rand}
\title{Generating alpha for GMM distribution}
\usage{
alpha_rand(n, m)
}
\arguments{
\item{n}{Number of alpha sets}

\item{m}{Number of random alphas}
}
\description{
Function generating alpha values of mixture distribution. Sum of alpha equals 1.
}
\examples{
alpha_rand(1,4)

}
