% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{calc_lLik2D}
\alias{calc_lLik2D}
\title{Log-Likelihood for 2D GMM}
\usage{
calc_lLik2D(X, Y, gmm)
}
\arguments{
\item{X}{x coordinates of histogram.}

\item{Y}{y coordinates of histogram.}

\item{gmm}{2D GMM fit, output of \code{\link{gaussian_mixture_2D}} function}
}
\description{
Function calculate log-likelihood of 2D Gaussian distribution
}
