% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{diag_init_2D}
\alias{diag_init_2D}
\title{Diagonal initialization of 2D GMM.}
\usage{
diag_init_2D(X, KS)
}
\arguments{
\item{X}{x coordinates of histogram.}

\item{K}{number of components.}
}
\description{
Function for generating initial conditions of 2D GMM model by diagonal.
}
