% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{diag_init_2D}
\alias{diag_init_2D}
\title{Supporting function for spliters.}
\usage{
diag_init_2D(X, KS)
}
\arguments{
\item{invec}{x coordinates of histogram.}

\item{yinwec}{y coordinates of histogram.}
}
\description{
Supporting function used in spliters \code{dyn_pr_split_w_aux} and \code{dyn_pr_split_w}.
}
