% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{DP_init_2D}
\alias{DP_init_2D}
\title{Dynamic programming split of aux.}
\usage{
DP_init_2D(X, Y, KS)
}
\arguments{
\item{X}{x coordinates of histogram.}

\item{ygreki}{y coordinates of histogram.}
}
\description{
Supporting function for dynamic programming in 1D decomposition.
}
