% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dsupport.R
\name{dyn_pr_split_w}
\alias{dyn_pr_split_w}
\title{Dynamic programming split.}
\usage{
dyn_pr_split_w(data, ygreki, K_gr, aux_mx)
}
\arguments{
\item{data}{x coordinates of histogram.}

\item{ygreki}{y coordinates of histogram.}

\item{K_gr}{Number of components.}

\item{aux_mx}{Return of function \code{dyn_pr_split_w_aux}.}
}
\value{
List of parameters.
}
\description{
Supporting function for dynamic programming in 1D decomposition.
}
