% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dsupport.R
\name{dyn_pr_split_w_aux}
\alias{dyn_pr_split_w_aux}
\title{Dynamic programming split of aux.}
\usage{
dyn_pr_split_w_aux(xhist, yhist)
}
\arguments{
\item{xhist}{x coordinates of histogram.}

\item{yhist}{y coordinates of histogram.}
}
\description{
Supporting function for dynamic programming in 1D decomposition.
}
