% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dplots.R
\name{ellips2D}
\alias{ellips2D}
\title{Ellipses for plot 2D GMM}
\usage{
ellips2D(center, covariance, cov_type, crit = 0.95)
}
\arguments{
\item{center}{Means of decomposition}

\item{covariance}{Covariances of each component}

\item{cov_type}{Type of covariance model same as in \code{\link{GMM_2D_opts}}. Possible "sphere","diag" or "full" (default).}

\item{crit}{Confidence interval level of ellipse. Default 0.95.}
}
\description{
Function for defining ellipses in 2D GMM plot by the confidence interval.
}
