% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_iter.R
\name{EM_iter}
\alias{EM_iter}
\title{Expectation–maximization algorithm for GMM}
\usage{
EM_iter(
  X,
  alpha,
  mu,
  sig,
  N,
  Y = NULL,
  change = Inf,
  max_iter = 5000,
  SW = 0.1,
  IC = "BIC"
)
}
\arguments{
\item{X}{Vector of data to decompose by GMM.}

\item{alpha}{Vector containing the weights (alpha) for each component in the statistical model.}

\item{mu}{Vector containing the means (mu) for each component in the statistical model.}

\item{sig}{Vector containing the standard deviation (sigma) for each component in the statistical model.}

\item{N}{Number of observations. Should be equal to \code{length(X)}
Applies only to binned data therefore the default is Y = NULL.}

\item{Y}{Vector of counts, should be the same length as "x".
Applies only to binned data therefore the default is Y = NULL.}

\item{change}{Stop of EM criterion (default Inf), value compared to following formula:
\deqn{\sum{(|\alpha - \alpha_{old})|} + \frac{\sum{(\frac{|\sigma^2 - \sigma^2_{old}|}{\sigma^2})}}{length(\alpha)}}}

\item{max_iter}{Maximum number of iterations of EM algorithm. By default it is \code{max_iter = 5000}}

\item{SW}{Minimum standard deviation of component (default 0.1). Values from 0 to 1 are treated according to their values, for SW >1 the following formula is applied:
\deqn{\frac{range(x)}{(SW*no.of.components))^2}}.}

\item{IC}{Information Criterion to select best number of components.
Possible "AIC","AICc", "BIC" (default), "ICL-BIC" or "LR".}
}
\value{
Returns a \code{list} of GMM parameter values that correspond to the local extremes for each component.\describe{
 \item{alpha}{Vector of optimal alpha (weights) values.}
 \item{mu}{Vector of optimal mu (means) values.}
 \item{sigma}{Vector of optimal sigma (standard devations) values.}
 \item{logL}{Log-likelihood value in local extreme.}
 \item{crit}{Value of the selected information criterion in local extreme of likelihood function.}
}
}
\description{
The function performs the EM algorithm to find the local maximum likelihood for the estimated Gaussian mixture parameters.
}
\seealso{
\code{\link{runGMM}} and \code{\link{gaussian_mixture_vector}}
}
