% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_iter_2D.R
\name{EM_iter_2D}
\alias{EM_iter_2D}
\title{Expectation–maximization algorithm for 2D data}
\usage{
EM_iter_2D(X, Y, init, opts)
}
\arguments{
\item{X}{matrix of data to decompose by GMM.}

\item{Y}{vector of counts, should be the same length as "X".
Applies only to binned data therefore the default is Y = NULL.}

\item{init}{vector of initial parameters for Gaussian components. See \code{\link{rand_init_2D}}, \code{\link{DP_init_2D}} or \code{\link{diag_init_2D}}.}

\item{opts}{parameters of run stored in \code{\link{GMM_2D_opts}} variable}
}
\value{
Function returns a \code{list} of GMM parameters for tested number of components: \describe{
 \item{alpha}{Weights (alpha) of each component}
 \item{center}{Means of decomposition}
 \item{covar}{Covariances of each component}
 \item{KS}{Number of components}
 \item{logL}{Log-likelihood value for the tested number of components}
 \item{IC}{Value of the selected information criterion which was used to calculate the tested number of components}
}
}
\description{
The function performs the EM algorithm to find the local maximum likelihood for the estimated Gaussian mixture parameters.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{runGMM2D}}
}
