% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dthresholds.R
\name{find_thr_by_dist}
\alias{find_thr_by_dist}
\title{Calculation of GMM thresholds from distribution}
\usage{
find_thr_by_dist(input)
}
\arguments{
\item{input}{output of \code{generate_dist} function. It is a list with following elements:\describe{
   \item{x}{Numeric vector with equaliy spread data of given precison}
   \item{dist}{Matrix with PDF of each GMM component and cumulative distribution}
}}
}
\value{
Return a vector of thresholds.
}
\description{
Function to calculate cutoffs between each component of mixture normal distributions using probability distribution function.
}
\examples{
data(example)
GModel<-data.frame(alpha=c(0.45,0.5,0.05),
                  mu=c(-14,-2,5),
                  sigma=c(2,4,1.5))
dist.plot<-generate_dist(example$Dist, GModel, 1e4)
thr <- find_thr_by_dist(dist.plot)

}
\seealso{
\code{\link{find_thr_by_params}} and \code{\link{generate_dist}}
}
