% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dthresholds.R
\name{find_thr_by_dist}
\alias{find_thr_by_dist}
\title{Thresholds estimations for 1D from GMM distribution}
\usage{
find_thr_by_dist(input, sigmas.dev = 2.5, GModel)
}
\arguments{
\item{input}{output of \code{\link{generate_dist}} function. It is a list with following elements:\describe{
   \item{x}{Numeric vector with equaliy spread data of given precison.}
   \item{dist}{Matrix with PDF of each GMM component and cumulative distribution.}
}}

\item{sigmas.dev}{Number of sigmas to secure thresholds on the ends of distributions. Equivalent to sigma.dev in merging GMMs.}

\item{GModel}{\code{data.frame} of GMM parameters i.e GModel$alpha, GModel$mu, GModel$sigma (correct \code{colnames} are obligatory).}
}
\value{
Return a vector of thresholds.
}
\description{
Function to calculate cutoffs between each component of mixture normal distributions using probability distribution function.
}
\examples{
\dontrun{
data(example)
GModel <- data.frame(alpha=c(0.45, 0.5, 0.05),
                    mu=c(-14, -2, 5),
                    sigma=c(2, 4, 1.5))
dist.plot <- generate_dist(example$Dist, GModel, 1e4)
thr <- find_thr_by_dist(dist.plot, 2.5, GModel)
}

}
\seealso{
\code{\link{runGMM}}
}
