% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_mixture_vector.R
\name{gaussian_mixture_vector}
\alias{gaussian_mixture_vector}
\title{Gaussian mixture decomposition for a vector of data}
\usage{
gaussian_mixture_vector(
  data,
  KS,
  Y = NULL,
  change = Inf,
  max_iter = 5000,
  SW = NULL,
  IC = "BIC",
  quick_stop = TRUE,
  signi = 0.05
)
}
\arguments{
\item{data}{Vector of data to decompose by GMM.}

\item{KS}{Maximum number of components to evaluate.}

\item{Y}{Vector of counts, should be the same length as "data".
Applies only to binned data therefore the default is Y = NULL.}

\item{change}{Stop of EM criterion (if < 1e-7). Default calculated as follow:
\deqn{\sum{(|alpha - alpha_{old})|} + \frac{\sum{(\frac{|sig2 - sig2_{old}|}{sig2})}}{length(alpha)}}}

\item{max_iter}{Maximum number of iterations of EM algorithm.}

\item{SW}{Minimum standard deviation of component.
Default set to: \deqn{\frac{range(x)}{(5*no.of.components))^2}}.}

\item{IC}{Information Criterion to select best number of components.
Possible "AIC","AICc", "BIC" (default), "ICL-BIC" or "LR".}

\item{quick_stop}{Logical value. Determines to stop the EM algorithm when adding
another component is no longer significant according to the Likelihood Ratio Test.
Used to speed up the function (Default is TRUE).}

\item{signi}{Significance level for Likelihood Ratio Test. By default is 0.05.}
}
\value{
Function returns a \code{list} of GMM parameters for the optimal number of components: \describe{
 \item{model}{A \code{data.frame} of model component parameters - mean values (mu), standard deviations (sigma)
 and weights (alpha) for each component. Output of \code{EM_iter}}
 \item{IC}{Value of the selected information criterion which was used to calculate the optimal number of components}
 \item{logL}{Log-likelihood value for the optimal number of components}
 \item{KS}{Optimal number of components}
}
}
\description{
Function to choose the optimal number of components of a mixture normal distributions, minimising the value of the information criterion.
}
\examples{
\dontrun{
data <- generate_norm1D(1000, alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3))
exp <- gaussian_mixture_vector(data$Dist, KS = 10, IC = "AIC", quick_stop = FALSE)
}

}
\seealso{
\code{\link{runGMM}} and \code{\link{EM_iter}}
}
