% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dgenerators.R
\name{generate_dist}
\alias{generate_dist}
\title{Generation of GMM data with high precision}
\usage{
generate_dist(data, GModel, precision)
}
\arguments{
\item{data}{Vector of original data}

\item{GModel}{\code{data.frame} of GMM parameters i.e GModel$alpha, GModel$mu, GModel$sigma (correct \code{colnames} are obligatory)}

\item{precision}{Precision of point linespacing}
}
\value{
List with following elements::\describe{
   \item{x}{Numeric vector with equaliy spread data of given precison}
   \item{dist}{Matrix with PDF of each GMM component and cumulative distribution}
}
}
\description{
Function to generate PDF of GMM distributions and its cumulative results with high lincespacing.
}
\seealso{
\code{\link{runGMM}} and \code{\link{generate_norm1D}}
}
