% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dist.R
\name{generate_dist}
\alias{generate_dist}
\title{Generation of GMM data with high precision}
\usage{
generate_dist(data, GModel, precision)
}
\arguments{
\item{data}{Vector of data}

\item{GModel}{List of GModel parameters i.e GModel$model$alpha, GModel$model$mu, GModel$model$sigma. Given as output of function gaussian_mixture_vector}

\item{precision}{Precision of point linespacing}
}
\value{
List with following elements::\describe{
   \item{x}{Numeric vector with equaliy spread data of given precison}
   \item{dist}{Matrix with PDF of each GMM component and cumulative distribution}
}
}
\description{
Function to generate GMM distribution data with their PDF and cumulative results.
}
\examples{
data<-generate_norm1D(1000, alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3))
GModel<-list(model=list(alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3)))
res<-generate_dist(data,GModel,1e4)

}
\seealso{
\code{\link{runGMM}}
}
