% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dgenerators.R
\name{generate_dset2D}
\alias{generate_dset2D}
\title{Create random 2D dataset}
\usage{
generate_dset2D(
  n = 1500,
  m = 1500,
  KS_range = 2:8,
  mu_range = c(-15, 15),
  sig_range = c(1, 5)
)
}
\description{
Create random 2D dataset
}
