% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_norm1D.R
\name{generate_norm1D}
\alias{generate_norm1D}
\title{QQplot for GMM decomposition}
\usage{
generate_norm1D(n, alpha, mu, sigma)
}
\arguments{
\item{n}{Number of points to generate}

\item{alpha}{Vector of alphas (weights) of each GMM component}

\item{mu}{Vector of means of each GMM component}

\item{sigma}{Vector of  sigmas of each GMM component}
}
\value{
List with following elements::\describe{
   \item{Dist}{Numeric vector with generated data}
   \item{Cls}{Numeric vector with labels of components to each point}
}
}
\description{
Generator of mixed-normal distribution
}
\details{
Function to generate mixed-normal distribution
}
\examples{
data<-generate_norm1D(1000, alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3))

}
