% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dgenerators.R
\name{generate_norm1D}
\alias{generate_norm1D}
\title{Generator of mixed-normal distributions}
\usage{
generate_norm1D(n, alpha, mu, sigma)
}
\arguments{
\item{n}{Number of points to generate}

\item{alpha}{Vector of alphas (weights) for each distribution}

\item{mu}{Vector of means for each distribution}

\item{sigma}{Vector of  sigmas for each distribution}
}
\value{
List with following elements::\describe{
   \item{Dist}{Numeric vector with generated data}
   \item{Cls}{Numeric vector with classification of each point to particular mixed distribution}
}
}
\description{
Generator of mixed-normal distribution with given model parameters for certain points number.
}
\examples{
\dontrun{
data<-generate_norm1D(1000, alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3))
}
}
