% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_2D_opts.R
\docType{data}
\name{GMM_2D_opts}
\alias{GMM_2D_opts}
\title{Default configuration for 2D Gaussian Mixture decomposition}
\format{
An object of class \code{gmm2_opts} of length 13.
}
\usage{
GMM_2D_opts
}
\arguments{
\item{eps_change}{Criterion for stopping the EM algorithm. By default it is \code{1e-4}}

\item{max_iter}{Maximum number of iterations of EM algorithm. By default it is \code{max_iter = 5000}}

\item{SW}{Regularizing coefficient for covariance (nie ja)}

\item{max_var_ratio}{Maximum dissimilarity between horizontal and vertical dispersion. By default it is \code{max_var_ratio = 5}}

\item{crit}{Information Criterion to select best number of components.
Possible "AIC","AICc", "BIC" (default) or "ICL-BIC".}

\item{res}{Parameter for modifying resolution (>1 less spots). **1** is as a default value. UŻYWAMY TEGO?!!!!!!!!!!!!!!!!!!!!!!!!}

\item{cov_type}{Type of covariance model. Possible "sphere","diag" or "full" (default).}

\item{init_nb}{Number of random initial conditions. By default it is \code{init_nb = 10}.}

\item{KS}{Maximum number of GMM components. By default it is \code{KS = 5}.}

\item{sig}{Logical value. Determines to stop the EM algorithm when adding another component is no longer significant according to the Likelihood Ratio Test. Used to speed up the function (Default is=FALSE).}

\item{D_thr}{Significance threshold for D statistic By default is 0.1 (Likelihood Ratio Test).}

\item{init_con}{Type of initial conditions. Could be "rand" (default),"DP" or "diag". See \code{\link{rand_init_2D}}, \code{\link{DP_init_2D}} or \code{\link{diag_init_2D}}.}

\item{fixed}{perform GMM only for number of components given in KS (defailt=FALSE).}
}
\description{
A list with parameters customizing a GMM_2D. Each component of the
list is an effective argument for \code{\link{runGMM2D}}.
}
\examples{
# display all default settings
GMM_2D_opts

# create a new settings object
custom.settings <- GMM_2D_opts
custom.settings$crit <- "BIC"
custom.settings

}
\keyword{datasets}
