% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_merge.R
\name{gmm_merge}
\alias{gmm_merge}
\title{Merging of overlapping components}
\usage{
gmm_merge(GModel, sigmas.dev = 2.5)
}
\arguments{
\item{GModel}{A \code{data.frame} of model component parameters - the rows are, in turn:  mean values (mu), standard deviations (sigma)
and weights (alpha) and each column corresponds to one component. Output of \code{EM_iter}.}

\item{sigmas.dev}{Parameter used to define close GMM components that needs to be merged. For each component, standard deviation is multiplied by \code{sigmas.dev} to estimate the distance from component mean.
All other components within this distance are merged. By default it is \code{sigmas.dev = 2.5}. When \code{sigmas.dev = 0} no components are merged.}
}
\value{
Function returns a \code{list} which contains: \describe{
  \item{model}{\code{data.frame} of the GMM decomposition parameters after merging.}
  \item{KS}{number of components after merging.}
}
}
\description{
Function merges distributions of overlapping components at distances defined in the argument \code{sigmas.dev}.
}
\seealso{
\code{\link{runGMM}}
}
\keyword{internal}
