% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dsupport.R
\name{my_qu_ix_w}
\alias{my_qu_ix_w}
\title{Supporting function for spliters}
\usage{
my_qu_ix_w(xinvec, yinvec, s_corr)
}
\arguments{
\item{xinvec}{x coordinates of histogram.}

\item{yinvec}{y coordinates of histogram.}
}
\description{
Supporting function used in spliters \code{dyn_pr_split_w_aux} and \code{dyn_pr_split_w}.
}
\keyword{internal}
