% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dsupport.R
\name{my_qu_ix_w}
\alias{my_qu_ix_w}
\title{Supporting function for spliters.}
\usage{
my_qu_ix_w(invec, yinwec)
}
\arguments{
\item{invec}{x coordinates of histogram.}

\item{yinwec}{y coordinates of histogram.}
}
\description{
Supporting function used in spliters \code{dyn_pr_split_w_aux} and \code{dyn_pr_split_w}.
}
