% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dsupport.R
\name{norm_pdf_2D}
\alias{norm_pdf_2D}
\title{Probability distribution of 2D normal distribution}
\usage{
norm_pdf_2D(x, center, covar)
}
\arguments{
\item{center}{centers of 2D distributions (means)}

\item{covar}{matrix of covariances}

\item{X}{matrix of 2D GMM data.}
}
\description{
Function for calculation PDF of 2D normal model.
}
\keyword{internal}
