% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dplots.R
\name{plot_gmm_1D}
\alias{plot_gmm_1D}
\title{Plot of GMM decomposition}
\usage{
plot_gmm_1D(X, dist, Y = NULL, threshold = NA, pal = NULL)
}
\arguments{
\item{X}{Vector of data}

\item{dist}{Output of \code{generate_dist} function. Its list of following elements\describe{
   \item{x}{Vector of generated data}
   \item{dist}{Matrix of pdf of for each generated mixture model. The last column is sum of all previous ones}
}}

\item{Y}{Vector of X counts (dedicated to binned data). Default=NULL}

\item{threshold}{Vector with GMM cutoffs}

\item{pal}{RColorBrewer palette name}
}
\description{
Function plot the decomposed distribution together with histogram of data. Moreover the cut-off are marked.
This plot is also return as regular output of \code{\link{runGMM}}.
}
\examples{
data(example)
GModel<-data.frame(alpha=c(0.45,0.5,0.05),
                  mu=c(-14,-2,5),
                  sigma=c(2,4,1.5))
dist.plot<-generate_dist(example$Dist, GModel, 1e4)
thr <- find_thr_by_params(GModel,dist.plot)
plot_gmm_1D(example$Dist, dist.plot,Y=NULL,threshold=thr, pal="Dark2")

}
\seealso{
\code{\link{runGMM}}
}
