% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gmm_1D.R
\name{plot_gmm_1D}
\alias{plot_gmm_1D}
\title{Plot of GMM decomposition for data vector}
\usage{
plot_gmm_1D(X, dist, Y = NULL, threshold = NA, pal = NULL)
}
\arguments{
\item{X}{Vector of data}

\item{dist}{Output of generate_dist function. Its list of following elements\describe{
   \item{x}{Vector of generated data}
   \item{dist}{Matrix of pdf of x for generated mixture model. The last column is sum of all previous ones.}
}}

\item{Y}{Vector of X counts (binned data). Default=NULL}

\item{threshold}{Vector with GMM cutoffs}

\item{pal}{RColorBrewer palette name}
}
\description{
Function plot the decomposed distribution together with histogram of data. Moreover the cut-off are marked.
}
\seealso{
\code{\link{runGMM}} and \code{\link{generate_dist}}
}
