% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dplots.R
\name{plot_gmm_2D_binned}
\alias{plot_gmm_2D_binned}
\title{Plot of GMM decomposition for 2D binned data}
\usage{
plot_gmm_2D_binned(X, Y, gmm, opts)
}
\arguments{
\item{X}{matrix of data to decompose by GMM.}

\item{Y}{vector of counts, should be the same length as "X".}

\item{gmm}{results of \code{\link{gaussian_mixture_2D}} decomposition}

\item{opts}{parameters of run saves in \code{\link{GMM_2D_opts}} variable}
}
\description{
Function plot the decomposed distribution together with histogram of data.
This plot is also return as regular output of \code{\link{runGMM}}.
}
\examples{
\dontrun{
}

}
