% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2Dplots.R
\name{plot_gmm_2D_orig}
\alias{plot_gmm_2D_orig}
\title{Plot of GMM decomposition for 2D data}
\usage{
plot_gmm_2D_orig(X, gmm, opts)
}
\arguments{
\item{X}{Matrix of data}

\item{gmm}{results of \code{\link{gaussian_mixture_2D}} decomposition}

\item{opts}{parameters of run stored in \code{\link{GMM_2D_opts}} variable}
}
\description{
Function plot the decomposed distribution together with histogram of data.
This plot is also return as regular output of \code{\link{runGMM}}.
}
\examples{
\dontrun{
}

}
