% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Dplots.R
\name{plot_QQplot}
\alias{plot_QQplot}
\title{QQplot of GMM decomposition for 1D data}
\usage{
plot_QQplot(X, alpha, mu, sigma)
}
\arguments{
\item{X}{Vector of 1D data for GMM decomposition.}

\item{GModel}{\code{data.frame} of GMM parameters i.e GModel$alpha, GModel$mu, GModel$sigma (correct \code{colnames} are obligatory).}
}
\description{
Function return ggplot object with fit diagnostic Quantile-Quantile plot for one normal distribution and fitted GMM.
This plot is also return as regular output of \code{\link{runGMM}}.
}
\examples{
\dontrun{
data(example)

alpha <- c(0.45, 0.5, 0.05)
mu <- c(-14, -2, 5)
sigma <- c(2, 4, 1.5)

plot_QQplot(example$Dist, alpha, mu, sigma)
}

}
\seealso{
\code{\link{runGMM}}
}
