% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QQplot.R
\name{plot_QQplot}
\alias{plot_QQplot}
\title{QQplot for GMM decomposition}
\usage{
plot_QQplot(data, GModel)
}
\arguments{
\item{data}{Vector of data}

\item{GModel}{List of GModel parameters i.e GModel$model$alpha, GModel$model$mu, GModel$model$sigma. Given as output of function gaussian_mixture_vector.}
}
\description{
Function return ggplot object with fit diagnostic Quantile-Quantile plot for one normal distribution and fitted GMM.
The QQplot is also return as regular output of runGMM.
}
\examples{
data<-generate_norm1D(1000, alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3))
GModel<-list(model=list(alpha=c(0.2,0.4,0.4), mu=c(-15,0,15), sigma=c(1,2,3)))
plot_QQplot(data$Dist,GModel)
}
\seealso{
\code{\link{runGMM}} and \code{\link{gaussian_mixture_vector}}
}
