% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMM2D.R
\name{runGMM2D}
\alias{runGMM2D}
\title{Function to perform Gaussian Mixture Model on 2D data}
\usage{
runGMM2D(X, Y = NULL, opts, plot = FALSE)
}
\arguments{
\item{X}{matrix of data to decompose by GMM.}

\item{Y}{vector of counts, should be the same length as "X".
Applies only to binned data therefore the default is Y = NULL.}

\item{opts}{parameters of run stored in \code{\link{GMM_2D_opts}} variable}

\item{plot}{logical value. If TRUE (default=FALSE), the GMM decomposition figure will be displayed.}
}
\value{
Function returns a \code{list} of GMM parameters for the optimal number of components: \describe{
 \item{model}{\describe{
    \item{alpha}{Weights (alpha) of each component}
    \item{center}{Means of decomposition}
    \item{covar}{Covariances of each component}
    \item{KS}{Optimal number of components}
    \item{logL}{Log-likelihood value for the optimal number of components}
    \item{IC}{Value of the selected information criterion which was used to calculate the optimal number of components}}}
 \item{fig}{Plot of decomposition}
}
}
\description{
Main function to perform GMM on 2D data. Function choose the optimal number of components of a 2D mixture normal distributions by minimising the value of the information criterion.
}
\examples{
\dontrun{
data(example2D)
opts<-GMM_2D_opts
opts$fixed<-T
opts$KS<-3
res<-runGMM2D(example2D[,1:2], example2D[,3], opts, plot=T)
}


}
